% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{Colorize numerical values}
\usage{
colorize(x, resolution = 4, pal = viridis::plasma)
}
\arguments{
\item{x}{Values to normalize}

\item{resolution}{Limit resolution for small values}

\item{pal}{Color palette (?viridis)}
}
\value{
Colorized values
}
\description{
Colorize numerical values
}
\examples{
colorize(c(1,2,3))

}
