% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Random values}
\format{
A numerical vector
\describe{
  \item{print(example)}{Vector elements}
}
}
\usage{
example
}
\description{
A vector of randomly chosen values.
}
\keyword{datasets}
