\name{phylo.dist}
\alias{phylo.dist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Cophenetic phylogenetic distance object }
\description{
  Wrapper for cophenetic returns a dist object sorted by species names
}
\usage{
phylo.dist(phylo)
}

\arguments{
  \item{phylo}{ phylo object }
}

\value{
 dist object
}

\author{ Steve Kembel <skembel@berkeley.edu> }
