% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgs-package.R
\docType{package}
\name{rpkgs-package}
\alias{rpkgs}
\alias{rpkgs-package}
\title{rpkgs: Just Trying Out Making A Package}
\description{
Just trying to get this package making to work.
}
\author{
\strong{Maintainer}: Octavia Dancu \email{octavia-maria.dancu@email.mcgill.ca} (YOUR-ORCID-ID)

}
\keyword{internal}
