\name{scanpkgs}
\alias{scanpkgs}
\docType{package}
\title{
\packageTitle{scanpkgs}
}
\description{
\packageDescription{scanpkgs}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{scanpkgs}

Maintainer: \packageMaintainer{scanpkgs}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = scanpkgs)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{scanpkgs}}
%\section{Documentation Index}{\packageIndices{scanpkgs}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
