% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize colors}
\usage{
visualize(cols, by = 100)
}
\arguments{
\item{cols}{An vector of colors}

\item{by}{Multiplier determining number of points}
}
\value{
Null
}
\description{
Visualize colors
}
\examples{
vals <- c(1,5,10,25)
cols <- colorize(vals)
visualize(cols)

}
